completed task 50

+++ task47/Program.cs
@@ -1,6 +1,6 @@
 ﻿double[,] CreateMatrixRndInt(int rows, int columns, double min, double max)
 {
-    int[,] matrix = new int[rows, columns];
+    double[,] matrix = new double[rows, columns];
     Random rnd = new Random();
 
     for (int i = 0; i < matrix.GetLength(0); i++)
